created Levels.js, added levels functions from Board.js to as level methods
