+++ lib/Board.js
@@ -1,6 +1,7 @@
 const Paddle = require('./Paddle.js');
 const Block = require('./Block.js');
 const Ball = require('./Ball.js')
+const Levels = require('./Levels.js')
 // import Paddle from './Paddle'
 let start = null;
 
@@ -10,9 +11,9 @@ class Board {
     this.width = 500
     this.height = 500
     this.paddle = new Paddle({x: 225, height: 10, width: 50})
-    this.block = level1()
-    this.block2 = level2()
-    this.block3 = winGame()
+    this.block =  new Levels().level1()
+    this.block2 = new Levels().level2()
+    this.block3 = new Levels().winGame()
     this.level = 1
     this.score = 0
     this.lives = 5
@@ -277,82 +278,82 @@ class Board {
 }
 
 
-function level1() {
-  let blockArray = []
-  let z = 1;
-  let c = -1
-
-  for (let i = 0; i < 60; i++) {
-    let x = (i % 10) * 50
-
-    if (i % 10 === 0) {
-      z++
-      c++
-    }
-    let y = z * 25
-    let blockColors = ['#bc3654', '#bc2e24', '#bc4b24', '#bcae24', '#24be24', '#2436bc']
-
-    blockArray.push(new Block({x, y, width: 50, height: 25, color: blockColors[c]}))
-
-  }
-  return blockArray
-}
-
-
-function level2() {
-  let blockArray2 = []
-  let z = 1;
-  let c = -1
-
-  for (let i = 0; i < 60; i++) {
-    let x = (i % 10) * 50
-
-    if (i % 10 === 0) {
-      z++
-      c++
-    }
-    let y = z * 25
-    let blockColors = ['white', 'green', 'red', 'orange', 'yellow', 'purple']
-
-    blockArray2.push(new Block({x, y, width: 50, height: 25, color: blockColors[c]}))
-  }
-  return blockArray2
-}
+// function level1() {
+//   let blockArray = []
+//   let z = 1;
+//   let c = -1
+//
+//   for (let i = 0; i < 60; i++) {
+//     let x = (i % 10) * 50
+//
+//     if (i % 10 === 0) {
+//       z++
+//       c++
+//     }
+//     let y = z * 25
+//     let blockColors = ['#bc3654', '#bc2e24', '#bc4b24', '#bcae24', '#24be24', '#2436bc']
+//
+//     blockArray.push(new Block({x, y, width: 50, height: 25, color: blockColors[c]}))
+//
+//   }
+//   return blockArray
+// }
 
 
+// function level2() {
+//   let blockArray2 = []
+//   let z = 1;
+//   let c = -1
+//
+//   for (let i = 0; i < 60; i++) {
+//     let x = (i % 10) * 50
+//
+//     if (i % 10 === 0) {
+//       z++
+//       c++
+//     }
+//     let y = z * 25
+//     let blockColors = ['white', 'green', 'red', 'orange', 'yellow', 'purple']
+//
+//     blockArray2.push(new Block({x, y, width: 50, height: 25, color: blockColors[c]}))
+//   }
+//   return blockArray2
+// }
 
-function winGame () {
-  let winArray = [
-    1, 0, 0, 0, 1, 2, 1, 0, 0, 1,
-    1, 0, 0, 0, 1, 2, 1, 0, 0, 1,
-    1, 0, 1, 0, 1, 2, 1, 1, 0, 1,
-    1, 0, 1, 0, 1, 2, 1, 1, 0, 1,
-    1, 0, 1, 0, 1, 2, 1, 1, 1, 1,
-    1, 0, 1, 0, 1, 2, 1, 0, 1, 1,
-    1, 1, 1, 1, 1, 2, 1, 0, 1, 1,
-    1, 1, 0, 1, 1, 2, 1, 0, 0, 1,
-    1, 0, 0, 0, 1, 2, 1, 0, 0, 1]
 
-  let z = 1;
 
-  for (let i = 0; i < 90; i++) {
-    let x = (i % 10) * 50
-
-    if (i % 10 === 0) {
-      z++
-    }
-    let y = z * 25
-
-    if (winArray[i] == 0) {
-      winArray.splice(i, 1, new Block({x, y, width: 50, height: 25, color: '#040204'}))
-    } else if (winArray[i] == 1) {
-      winArray.splice(i, 1, new Block({x, y, width: 50, height: 25, color: '#bc3654'}))
-    } else if (winArray[i] == 2) {
-      winArray.splice(i, 1, new Block({x, y, width: 50, height: 25, color: '#bc4b24'}))
-    }
-  }
-  return winArray
-}
+// function winGame () {
+//   let winArray = [
+//     1, 0, 0, 0, 1, 2, 1, 0, 0, 1,
+//     1, 0, 0, 0, 1, 2, 1, 0, 0, 1,
+//     1, 0, 1, 0, 1, 2, 1, 1, 0, 1,
+//     1, 0, 1, 0, 1, 2, 1, 1, 0, 1,
+//     1, 0, 1, 0, 1, 2, 1, 1, 1, 1,
+//     1, 0, 1, 0, 1, 2, 1, 0, 1, 1,
+//     1, 1, 1, 1, 1, 2, 1, 0, 1, 1,
+//     1, 1, 0, 1, 1, 2, 1, 0, 0, 1,
+//     1, 0, 0, 0, 1, 2, 1, 0, 0, 1]
+//
+//   let z = 1;
+//
+//   for (let i = 0; i < 90; i++) {
+//     let x = (i % 10) * 50
+//
+//     if (i % 10 === 0) {
+//       z++
+//     }
+//     let y = z * 25
+//
+//     if (winArray[i] == 0) {
+//       winArray.splice(i, 1, new Block({x, y, width: 50, height: 25, color: '#040204'}))
+//     } else if (winArray[i] == 1) {
+//       winArray.splice(i, 1, new Block({x, y, width: 50, height: 25, color: '#bc3654'}))
+//     } else if (winArray[i] == 2) {
+//       winArray.splice(i, 1, new Block({x, y, width: 50, height: 25, color: '#bc4b24'}))
+//     }
+//   }
+//   return winArray
+// }
 
 // Board.prototype.paddleCollision = function () {
 //   if (this.ball.x < this.paddle.x + this.paddle.width &&
